requirements.txt
numpy==1.23.5
pillow==9.3.0
tensorflow==2.11.0
keras==2.11.0
matplotlib==3.6.2
opencv-python==4.6.0.66
scikit-image==0.19.3
scikit-learn==1.1.3
albumentations==1.2.1
tqdm==4.64.1
